+++ expensify/components/ExpensesDashBoard.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ExpensesDashBoard = () => {
+    return (
+        <div>
+            <h1>This is Expenses Dash Board</h1>
+        </div>
+    )
+}
+
+export { ExpensesDashBoard }
