+++ expensify/components/NotFoundPage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const NotFound = () => {
+    return (
+        <div>
+            <h1>Page Not Found Error...</h1>
+        </div>
+    )
+}
+
+export { NotFound }
