+++ expensify/components/addExpense.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const addExpense = () => {
+    return (
+        <div>
+            <h1>This is Add Expense Component</h1>
+        </div>
+    )
+}
+
+export { addExpense }
