+++ expensify/components/editExpense.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const editExpense = () => {
+    return (
+        <div>
+            <h1>Edit Expense page....</h1>
+        </div>
+    )
+}
+
+export { editExpense }
