+++ expensify/components/help.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const HelpPage = () => {
+    return (
+        <div>
+            <h1>This is Help Page...</h1>
+        </div>
+    )
+}
+
+export { HelpPage };
